refactor: splitt scenarios.ts i domenefiler — campaign, scenario, encounter, participant, template (#47)
scenarios.ts (1225 linjer) er erstattet med en 12-linjer orchestrator som registrerer:
- templateRoutes.ts  — /templates/*, /entities/*
- campaignRoutes.ts  — /campaigns/*, /campaign-assets/*
- scenarioRoutes.ts  — /scenarios/joinable, /scenarios/:id CRUD, live-state, events
- encounterRoutes.ts — /scenarios/:id/encounters, /encounters/:id
- participantRoutes.ts — /scenarios/:id/participants/*

Alle eksisterende tester passerer (21/21 grønne).
