+++ apps/api/src/routes/scenarios/encounterRoutes.ts
@@ -1,0 +1,181 @@
+import type { FastifyPluginAsync } from "fastify";
+
+import { EncounterService, ScenarioService } from "@glantri/database";
+import { encounterSessionSchema } from "@glantri/domain";
+
+import { requireAdminUser, requireAuthenticatedUser } from "../../lib/sessionAuth";
+import { parseBodyObject, parseId } from "./parsing";
+import { resolveScenarioWorkspaceAccess } from "./access";
+
+const encounterService = new EncounterService();
+const scenarioService = new ScenarioService();
+
+export const encounterRoutes: FastifyPluginAsync = async (app) => {
+  app.get("/scenarios/:scenarioId/encounters", async (request, reply) => {
+    const user = await requireAuthenticatedUser(request, reply);
+
+    if (!user) {
+      return;
+    }
+
+    try {
+      const scenarioId = parseId(request.params, "scenarioId", "Scenario id");
+      const access = await resolveScenarioWorkspaceAccess(scenarioService, {
+        scenarioId,
+        userId: user.id,
+        userRoles: user.roles
+      });
+
+      if (!access) {
+        return reply.code(404).send({
+          error: "Scenario not found."
+        });
+      }
+
+      const encounters = await encounterService.listEncountersByScenario(scenarioId);
+
+      return { encounters };
+    } catch (error) {
+      return reply.code(400).send({
+        error: error instanceof Error ? error.message : "Unable to list encounters."
+      });
+    }
+  });
+
+  app.post("/scenarios/:scenarioId/encounters", async (request, reply) => {
+    const user = await requireAdminUser(request, reply);
+
+    if (!user) {
+      return;
+    }
+
+    try {
+      const scenarioId = parseId(request.params, "scenarioId", "Scenario id");
+      const access = await resolveScenarioWorkspaceAccess(scenarioService, {
+        scenarioId,
+        userId: user.id,
+        userRoles: user.roles
+      });
+
+      if (!access || access.mode !== "gm") {
+        return reply.code(404).send({
+          error: "Scenario not found."
+        });
+      }
+
+      const body = parseBodyObject(request.body, "Encounter payload");
+      const encounter = await encounterService.createEncounter({
+        createdByUserId: user.id,
+        session: encounterSessionSchema.parse(body.session)
+      });
+
+      await scenarioService.recordScenarioEvent({
+        actorUserId: user.id,
+        eventType: "encounter_created",
+        payload: {
+          encounterId: encounter.id,
+          kind: encounter.kind,
+          status: encounter.status
+        },
+        scenarioId,
+        summary: `Created ${encounter.kind} encounter ${encounter.title}.`
+      });
+
+      return { encounter };
+    } catch (error) {
+      return reply.code(400).send({
+        error: error instanceof Error ? error.message : "Unable to create encounter."
+      });
+    }
+  });
+
+  app.get("/encounters/:encounterId", async (request, reply) => {
+    const user = await requireAuthenticatedUser(request, reply);
+
+    if (!user) {
+      return;
+    }
+
+    try {
+      const encounterId = parseId(request.params, "encounterId", "Encounter id");
+      const encounter = await encounterService.getEncounterById(encounterId);
+
+      if (!encounter?.scenarioId) {
+        return reply.code(404).send({
+          error: "Encounter not found."
+        });
+      }
+
+      const access = await resolveScenarioWorkspaceAccess(scenarioService, {
+        scenarioId: encounter.scenarioId,
+        userId: user.id,
+        userRoles: user.roles
+      });
+
+      if (!access) {
+        return reply.code(404).send({
+          error: "Encounter not found."
+        });
+      }
+
+      return { encounter };
+    } catch (error) {
+      return reply.code(400).send({
+        error: error instanceof Error ? error.message : "Unable to load encounter."
+      });
+    }
+  });
+
+  app.put("/encounters/:encounterId", async (request, reply) => {
+    const user = await requireAdminUser(request, reply);
+
+    if (!user) {
+      return;
+    }
+
+    try {
+      const encounterId = parseId(request.params, "encounterId", "Encounter id");
+      const existingEncounter = await encounterService.getEncounterById(encounterId);
+
+      if (!existingEncounter?.scenarioId) {
+        return reply.code(404).send({
+          error: "Encounter not found."
+        });
+      }
+
+      const access = await resolveScenarioWorkspaceAccess(scenarioService, {
+        scenarioId: existingEncounter.scenarioId,
+        userId: user.id,
+        userRoles: user.roles
+      });
+
+      if (!access || access.mode !== "gm") {
+        return reply.code(404).send({
+          error: "Encounter not found."
+        });
+      }
+
+      const body = parseBodyObject(request.body, "Encounter payload");
+      const encounterSession = encounterSessionSchema.parse(body.session);
+      const encounter = await encounterService.updateEncounter(encounterSession);
+
+      await scenarioService.recordScenarioEvent({
+        actorUserId: user.id,
+        eventType: "encounter_updated",
+        payload: {
+          encounterId: encounter.id,
+          participantCount: encounter.participants.length,
+          status: encounter.status
+        },
+        scenarioId: existingEncounter.scenarioId,
+        summary: `Updated encounter ${encounter.title}.`
+      });
+
+      return { encounter };
+    } catch (error) {
+      return reply.code(400).send({
+        error: error instanceof Error ? error.message : "Unable to update encounter."
+      });
+    }
+  });
+};
